Finish Mandelbrot and Julia generator functions
Finished the code for generating and displaying both the Mandelbrot and Julia fractal generator functions. Currently using unvectorised operations, may attempt to implement that in the future.

Next steps are to implement a UI to allow the user to continuously generate different Julia sets as well as adjust values in an easier fashion